# improve performance of GdiAlphaBitmapSurface::Blend